refactorizacion y cambios en entidades
