+++ proyecto-tad/src/main/java/upo/tad/g11/proyecto/tad/MyUI.java
@@ -1,4 +1,4 @@
-package upo.tad.witzl.proyecto.tad;
+package upo.tad.g11.proyecto.tad;
 
 import com.mongodb.MongoClient;
 import javax.servlet.annotation.WebServlet;
