+++ proyecto-tad/src/main/java/upo/tad/g11/proyecto/tad/controller/Controlador.java
@@ -1,7 +1,7 @@
 /*
  * Esta interfaz define el contenido minimo del controlador
  */
-package upo.tad.witzl.proyecto.tad.Controller;
+package upo.tad.g11.proyecto.tad.controller;
 
 /**
  *
