+++ proyecto-tad/src/main/java/upo/tad/g11/proyecto/tad/model/Connection.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package upo.tad.witzl.proyecto.tad.model;
+package upo.tad.g11.proyecto.tad.model;
 
 import com.mongodb.MongoClient;
 import com.mongodb.client.MongoDatabase;
