+++ proyecto-tad/src/main/java/upo/tad/g11/proyecto/tad/model/DAO.java
@@ -2,13 +2,14 @@
  * En esta interfaz se detallan las acciones minimas que debe realizar un objeto DAO
  * 
  */
-package upo.tad.witzl.proyecto.tad.model;
+package upo.tad.g11.proyecto.tad.model;
 
 import java.util.List;
 
 /**
  *
  * @author Amalio
+ * @param <T>
  */
 public interface DAO<T> {
